[r] Move VerticalStepperModule to /account/
More specifically, /account/trial/
