+++ apps/web/src/app/accounts/trial-initiation/vertical-stepper/vertical-step.component.html
