+++ apps/web/src/app/accounts/trial-initiation/vertical-stepper/vertical-stepper.component.ts
